feat(nn): quantization enabled, not sure of correctness

+++ src/engine.rs
@@ -5,7 +5,7 @@ pub struct Engine {
     pub tt: TranspositionTable,
     pub history: [[[i32; 64]; 64]; 2],
     pub killers: [[Move; 2]; MAX_PLY],
-    pub accumulators: [[f32; HL1]; MAX_PLY],
+    pub accumulators: [[i32; HL1]; MAX_PLY],
 }
 
 impl Engine {
@@ -15,7 +15,7 @@ impl Engine {
             tt: TranspositionTable::new(16),
             history: [[[0; 64]; 64]; 2],
             killers: [[Move::NULL; 2]; MAX_PLY],
-            accumulators: [[0f32; HL1]; MAX_PLY],
+            accumulators: [[0i32; HL1]; MAX_PLY],
         }
     }
 }
